+++ src/main/kotlin/app/ubie/spannerkase/internal/MigrationDataFile.kt
@@ -1,8 +1,8 @@
-package app.ubie.spannerkase
+package app.ubie.spannerkase.internal
 
 import java.io.File
 
-class MigrationDataFile(
+internal class MigrationDataFile(
     private val file: File
 ) : MigrationData() {
     override val sql: String by lazy {
